+++ test/Enum.jl
@@ -8,19 +8,18 @@
         renamefields!(b, StandardEnum, [:se_a => "SeA", :se_b => "SeB", :se_c => "SeC"])
         sb = Reflect.StringLayouts(b)
 
-
         sb[StandardEnum] === """#[repr(i32)]
-        #[jlrs(julia_type = "Main.StandardEnum")]
+        #[jlrs(julia_type = "StandardEnum")]
         #[derive(Copy, Clone, Debug, PartialEq, Enum, Unbox, IntoJulia, ConstructType, IsBits, Typecheck, ValidField, ValidLayout, CCallArg, CCallReturn)]
         enum StandardEnum {
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.se_a")]
+            #[jlrs(julia_enum_variant = "se_a")]
             SeA = 1,
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.se_b")]
+            #[jlrs(julia_enum_variant = "se_b")]
             SeB = 2,
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.se_c")]
+            #[jlrs(julia_enum_variant = "se_c")]
             SeC = 3,
         }"""
 
@@ -31,17 +30,17 @@
         sb = Reflect.StringLayouts(b)
 
         sb[OtherTypeEnum] === """#[repr(i16)]
-        #[jlrs(julia_type = "Main.OtherTypeEnum")]
+        #[jlrs(julia_type = "OtherTypeEnum")]
         #[derive(Copy, Clone, Debug, PartialEq, Enum, Unbox, IntoJulia, ConstructType, IsBits, Typecheck, ValidField, ValidLayout, CCallArg, CCallReturn)]
         enum OtherTypeEnum {
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.ote_a")]
+            #[jlrs(julia_enum_variant = "ote_a")]
             ote_a = 1,
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.ote_b")]
+            #[jlrs(julia_enum_variant = "ote_b")]
             ote_b = -1,
             #[allow(non_camel_case_types)]
-            #[jlrs(julia_enum_variant = "Main.ote_c")]
+            #[jlrs(julia_enum_variant = "ote_c")]
             ote_c = 0,
         }"""
     end