+++ test/ZeroSized.jl
@@ -9,7 +9,7 @@ struct TypedEmpty{T} end
 
         sb[Empty] === """#[repr(C)]
         #[derive(Clone, Debug, Unbox, ValidLayout, Typecheck, IntoJulia, ValidField, IsBits, ConstructType)]
-        #[jlrs(julia_type = "Main.Empty", zero_sized_type)]
+        #[jlrs(julia_type = "Empty", zero_sized_type)]
         pub struct Empty {
         }"""
     end
@@ -22,12 +22,12 @@ end
 
         sb[Reflect.basetype(TypedEmpty)] === """#[repr(C)]
         #[derive(Clone, Debug, Unbox, ValidLayout, Typecheck, ValidField, IsBits)]
-        #[jlrs(julia_type = "Main.TypedEmpty")]
+        #[jlrs(julia_type = "TypedEmpty")]
         pub struct TypedEmpty {
         }
 
         #[derive(ConstructType, HasLayout)]
-        #[jlrs(julia_type = "Main.TypedEmpty", constructor_for = "TypedEmpty", scope_lifetime = false, data_lifetime = false, layout_params = [], elided_params = ["T"], all_params = ["T"])]
+        #[jlrs(julia_type = "TypedEmpty", constructor_for = "TypedEmpty", scope_lifetime = false, data_lifetime = false, layout_params = [], elided_params = ["T"], all_params = ["T"])]
         pub struct TypedEmptyTypeConstructor<T> {
             _t: ::std::marker::PhantomData<T>,
         }"""